improve GetHighCalorieMealsUseCase by Refactor class name and make it default and put 700 as parameter for suggestMeal function
